meta: WIP mail-link-to-pr.sh
This is just a WIP sketch.
The idea is to take a link to a patch series on lore.kernel.org and
convert it into a PR on this repo on github.

The PR is only used for building CI and reporting results.
No comments should be done on the PR as that should take place on the list.

Right now the titles etc are hard coded.

Signed-off-by: Bill Mills <bill.mills@linaro.org>