jailhouse-images Linux base for v5.10.31
The jailhouse-images repo [1] is an ISAR build system to build demo images
of Jailhouse.  As of 2023-03-18 the master branch builds a Linux it calls
v5.10.31.

It downloads a tar file of the kernel source from its own github. [2]
The SHA used [3] is the current commit on branch jailhouse-enabling/5.10 [4]

That branch is not a rebased branch so here is all the diffs between
stock v5.10.31 and the commit id above as one big patch.

The build system then applies more patches.  I have not included those
here but they are clearly visible in the meta data. [5]

I am  publishing this here for my own sanity and notes.

[1] https://github.com/siemens/jailhouse-images
[2] https://github.com/siemens/jailhouse-images/blob/master/recipes-kernel/linux/linux-jailhouse_5.10.inc#L17
[3] https://github.com/siemens/linux/commits/eb6927f7eea77f823b96c0c22ad9d4a2d7ffdfce
[4] https://github.com/siemens/linux/commits/jailhouse-enabling/5.10
[5] https://github.com/siemens/jailhouse-images/tree/master/recipes-kernel/linux/files

Signed-off-by: Bill Mills <bill.mills@linaro.org>