arm64: dts: zynqmp: Build and install the Kria KV overlays
In order to be installed the dtbo needs to be added to dtb-y.

In order for the overlay apply test to be run, it needs to be added to
dtb-y.  Right now there does not seem to be a way to test apply the
dtbo without the result being installed.

This Makefile also mentions the KR overlay but it is not in upstream kernel
yet so not adding those.  (But not removing from Makefile either.)

Signed-off-by: Bill Mills <bill.mills@linaro.org>