pinctrl: stmfx: Manage irq affinity settings
Trying to set the affinity of the interrupt associated to stmfx
pinctrl results in a write error because stmfx pinctrl was not
managing the set of the irq smp affinity. It is resolved by
adding the .irq_set_affinity to the stmfx pinctrl irq_chip struct and
fill it with a function to manage affinity. The function return an
error explaining to set the affinity of the gpio interrupt
provider to change stmfx pinctrl smp_affinity.

Change-Id: I584c14351bbc1541091f3c8006d6cb0d7f87ddfc
Signed-off-by: Cheick Traore <cheick.traore-ext@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/401706
Reviewed-by: Antonio Maria BORNEO <antonio.borneo@foss.st.com>
Domain-Review: Antonio Maria BORNEO <antonio.borneo@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>