firmware: arm_ffa: Extend interfaces to support framework notifications
Currentlt the existing notification interfaces(notify_* APIs) doesn't
support the framework notifications defined in the FF-A specification.

In order to support the same, extend the interfaces so that the
framework notifications can be indicated explicitly. This change
doesn't add any support for the framework notifications itself.

So, it is not functional change to any existing users of the interface
as they all are not using the framework notifications. However, if a
framework notification is requested, it will return as not supported
after this change.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>