kconfig: check-in openamp-ci kconfig fragments
For ease of use by kernel developers, check in the effective kconfig
fragments uses for arm64 and arm32 builds in the openamp CI builds.

These fragments are the result of multiple fragments from the OE meta data
but combine them here for ease of use.

I have also checked in the script I use to make them as documentation.
You should not need to run it.

Use them like this:
for arm64:
    make defconfig openamp-ci.config
for arm32:
    make multi_v7_defconfig openamp-ci.config

Signed-off-by: Bill Mills <bill.mills@linaro.org>