virtio_msg: dev_id in the message is the ID on the bus
The dev_id field in the message is the bus level ID of this device.
This ID is defined by the device side or VMM it is not assigned by
our driver side.  It would be similar to the slot number in pci.

The dev_id field must already be populated at register time.

dev_id is not the device type as previously coded.  A given bus can have
multiple virtio-net devices each with its own unique dev_id and each
would have its device type of 0001.

Signed-off-by: Bill Mills <bill.mills@linaro.org>