virtio_msg_amp: simplify and fix timeout wait
I was not really handling interruptions correctly anyway so use a simple
uninterruptible wait.

Somewhere along the way the polarity of the result got swapped but true
was always supposed to be OK and false was timeout or other error.

Increase timeout value so we can really see if it delays.

Tested timeout case by killing the device side after communication started.

Signed-off-by: Bill Mills <bill.mills@linaro.org>