virtio-msg: ffa: Add passive-mode host support
Allow the FF-A virtio-msg bus to run in “passive” mode when probed on
the host side (vm_id != 1). In this mode the backend only services
requests initiated elsewhere, so it routes FF-A events straight to the
virtio-msg userspace device instead of spinning up virtio devices.

Hook the FF-A transport into `virtio_msg_user` so that userspace can
read/write messages while the kernel thread remains idle, and make
probing/removal paths aware of the passive lifecycle.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>